add debug mode update timestamps to use the format defined in global.psm1 update date comparaison to be between date objects instead of strings. unhide hidden folders in VSC rollback changes to task registration.

+++ functions/process/Get-Lock.psm1
@@ -3,8 +3,12 @@ function Get-Lock {
   [OutputType([boolean])]
   param (
   )
+  if ($Global.Debug){
+    Write-ScriptMsg "Debug Mode, skipping Lock check..."
+    return $false
+  }
   if ((Test-Path -Path ".\servers\$($Server.Name).LOCK" -PathType "Leaf" -ErrorAction SilentlyContinue)) {
-    $TimeStamp = Get-IniValue -file ".\servers\$($Server.Name).LOCK" -category "Lock" -key "TimeStamp"
+    $TimeStamp = [datetime]::ParseExact((Get-IniValue -file ".\servers\$($Server.Name).LOCK" -category "Lock" -key "TimeStamp"), $Global.DateTimeFormat, $null)
     if ($TimeStamp -lt (Get-Date).AddMinutes(-$Global.LockTimeout)) {
       Write-ScriptMsg "Process Lock is too old, removing..."
       $null = Remove-Item -Path ".\servers\$($Server.Name).LOCK" -Force -ErrorAction SilentlyContinue
